syn: Move types out of `acd.cc` to `acd.h`
